python/pyproject.toml
[tool.poetry]
name = "python-api-samples"
version = "0.1.0"
description = ""
authors = ["Eles <1494136313@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.4"
black = "^23.12.1"
google-auth-oauthlib = "^1.2.0"
google-api-python-client = "^2.114.0"
oauthlib = "^3.2.2"
google-auth = "^2.26.2"
google-auth-httplib2 = "^0.2.0"
pytest-asyncio = "^0.23.3"
dependency-injector = "^4.41.0"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -s"
testpaths = [
    "tests",
    "integration",
]